+++ source/plugins/device.wyapp.websocket/index.js
@@ -119,6 +119,7 @@ export function setup (options, imports, register)
 
 	workspace.registerMenuItem('DEVICE_WYAPP_WEBSOCKET_SET_USER_ID', 90, async () => {
 		let settoken = await workspace.showPrompt ('DEVICE_WYAPP_WEBSOCKET_SET_USER_ID_TITLE', 'DEVICE_WYAPP_WEBSOCKET_SET_USER_ID_TEXT', token);
+		if (settoken && validator.isUUID (settoken)) settoken = settoken.toLowerCase ();
 		if (settoken && token !== settoken)
 		{
 			if (validator.isUUID (settoken))